requirements.txt
streamlit>=1.32.0
extra-streamlit-components>=0.1.71
plotly>=5.20.0
db-dtypes>=1.2.0
google-cloud-bigquery-storage>=2.24.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-cloud-bigquery>=3.17.0
python-dotenv>=1.0.0
requests>=2.31.0
